backend/requirements.txt
Django>=4.2.0
djangorestframework>=3.14.0
djangorestframework-simplejwt>=5.3.0
django-cors-headers>=4.0.0
python-dotenv>=1.0.0
psycopg2-binary>=2.9.0
gunicorn>=21.2.0

# Redis and Celery
redis>=5.0.0
celery>=5.3.0
django-redis>=5.4.0

# Rate limiting
django-ratelimit>=4.1.0

twilio>=8.10.0

langchain>=1.0.0
langchain-core>=1.0.0
langchain-community>=0.4.0
langchain-groq>=0.2.0
neo4j>=5.0.0

PyPDF2>=3.0.0
python-docx>=1.0.0
Pillow>=10.0.0
pytesseract>=0.3.10

# Vector Database
pymilvus>=2.3.0
sentence-transformers>=2.2.0

# Performance monitoring
psutil>=5.9.0

# Additional dependencies
groq>=0.4.0
python-magic>=0.4.27